# Conflicts:
#	jvm-core/src/main/java/indi/pancras/jvm/instruction/references/Invokespecial.java

+++ jvm-core/src/main/java/indi/pancras/jvm/instruction/conversion/I2L.java
@@ -18,6 +18,6 @@ public class I2L extends BaseNoOperands {
     @Override
     public void execute(Frame frame) {
         OperandStack operandStack = frame.getOperandStack();
-        operandStack.pushLong((long) operandStack.popInt());
+        operandStack.pushLong(operandStack.popInt());
     }
 }
